% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{circle_projector}
\alias{circle_projector}
\title{project points to a circle}
\usage{
circle_projector(
  x,
  y,
  x0 = 0,
  y0 = 0,
  r = 1,
  source = c("x", "y", "-x", "-y", "center")
)
}
\arguments{
\item{x, y}{points to be projected}

\item{x0, y0}{circle center}

\item{r}{radius of circle}

\item{source}{projection type.}
}
\value{
data.frame of coordinates of projected points
}
\description{
project points to a circle
}
\details{
Here, for sources of x or y, projected points will be at the right or up half circle;
for sources of -x or -y, projected points will be at the left or down half circle
}
\keyword{internal}
