% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosslink.R
\name{cl_active}
\alias{cl_active}
\alias{cl_active<-}
\title{Retrieve and Set active layout}
\usage{
cl_active(object, ret.data = F)

cl_active(object) <- value
}
\arguments{
\item{object}{a CrossLink object}

\item{ret.data}{return layout data}

\item{value}{the name of a layout to be set as active}
}
\value{
name or data of the active layout

an updated CrossLink object
}
\description{
Retrieve and Set active layout
}
\examples{


}
