% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosslink.R
\name{cl_align}
\alias{cl_align}
\title{Align crosses in different layouts}
\usage{
cl_align(
  object,
  crosses.1,
  crosses.2,
  align.x = F,
  align.y = F,
  anchor.1 = c(0.5, 0.5),
  anchor.2 = c(0.5, 0.5),
  layout = NULL
)
}
\arguments{
\item{object}{CrossLink object}

\item{crosses.1}{crosses in group 1}

\item{crosses.2}{crosses in group 2}

\item{align.x}{align both groups in x}

\item{align.y}{align both groups in y}

\item{anchor.1}{relative coordinates of the anchor in group 1 by which to align}

\item{anchor.2}{relative coordinates of the anchor in group 2 by which to align}

\item{layout}{which layout coordinates to be used}
}
\value{

}
\description{
Align crosses in different layouts
}
\examples{

}
