% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosslink.R
\name{cl_annotation}
\alias{cl_annotation}
\title{Helper for annotation in cl_plot}
\usage{
cl_annotation(
  top = NULL,
  top.by = NULL,
  top.height = 1,
  bottom = NULL,
  bottom.by = NULL,
  bottom.height = 1,
  left = NULL,
  left.by = NULL,
  left.width = 1,
  right = NULL,
  right.by = NULL,
  righ.width = 1
)
}
\arguments{
\item{top, bottom, left, right}{ggplot object}

\item{top.by, bottom.by, left.by, right.by}{name of cross by which to align ggplot}

\item{top.width, bottom.width, left.width, righ.width}{width of ggplot to be aligned, relative value to main plot.}
}
\value{
a list of arguments
}
\description{
Helper for annotation in cl_plot
}
\examples{

}
