% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosslink.R
\name{cl_layouts}
\alias{cl_layouts}
\title{Retrieve available layout names}
\usage{
cl_layouts(object)
}
\arguments{
\item{object}{a CrossLink object}
}
\value{
names of available layouts
}
\description{
Retrieve available layout names
}
\examples{

}
