% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosslink.R
\name{cl_plot}
\alias{cl_plot}
\title{Plot a CrossLink object}
\usage{
cl_plot(
  object,
  layout = NULL,
  link = NULL,
  cross = NULL,
  label = NULL,
  header = NULL,
  ...
)
}
\arguments{
\item{object}{a CrossLink object}

\item{layout}{the layout to be plot. Set NULL to plot current active layout}

\item{link}{a named list of arguments for links. see \code{\link[ggplot2:geom_segment]{ggplot2::geom_segment()}}. Set NULL to use default settings, or Set NA to not show.}

\item{cross}{a named list of arguments for crosses. see \code{\link[ggplot2:geom_point]{ggplot2::geom_point()}}. Set NULL to use default settings, or Set NA to not show.}

\item{label}{a named list of arguments for node labels. see \code{\link[ggplot2:geom_text]{ggplot2::geom_text()}}. Set NULL to use default settings, or Set NA to not show.}

\item{header}{a named list of arguments for headers. see \code{\link[ggplot2:geom_text]{ggplot2::geom_text()}}. Set NULL to use default settings, or Set NA to not show.}

\item{...}{other arguments passed to ???}
}
\value{
a ggplot2 object
}
\description{
Plot a CrossLink object
}
\details{
For link, cross, label, header: "geom" is restricted for setting geom function, such as list(geom = "arc") for link.
}
\examples{

}
