% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosslink.R
\name{cl_plot2}
\alias{cl_plot2}
\title{Plot a CrossLink object}
\usage{
cl_plot2(p, object, layout = NULL, annotation = cl_annotation())
}
\arguments{
\item{p}{a ggplot2 object for crosslink plot}

\item{object}{a CrossLink object}

\item{layout}{the layout to be plot. Set NULL to plot current active layout}

\item{annotation}{generated by cl_annotation()}
}
\value{
a ggplot2 object
}
\description{
Plot a CrossLink object
}
