% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosslink.R
\name{cl_rel2abs}
\alias{cl_rel2abs}
\title{Convert relative coordinates}
\usage{
cl_rel2abs(object, x = 0, y = 0, crosses = NULL, by.canvas = F, layout = NULL)
}
\arguments{
\item{object}{a CrossLink object}

\item{x, y}{relative coordinates}

\item{crosses}{relative about which cross}

\item{by.canvas}{relative about whole canvas}

\item{layout}{layout name to retrieve coordinates}
}
\value{
data.frame
}
\description{
Convert relative coordinates
}
\examples{

}
