% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosslink.R
\name{cl_void}
\alias{cl_void}
\title{a theme for crosslink plot}
\usage{
cl_void(gg, th = theme())
}
\arguments{
\item{gg}{a ggplot object}

\item{th}{a value returned by theme()}
}
\value{
an updated ggplot object
}
\description{
a theme for crosslink plot
}
\examples{

}
