% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{coerce_x_len}
\alias{coerce_x_len}
\title{use a vector repeatedly to generate a new vector of the specified length}
\usage{
coerce_x_len(x, n, unlist = T)
}
\arguments{
\item{x}{a vector to be used repeatedly}

\item{n}{length of the new generated vector}

\item{unlist}{unlist \code{x} before repeatedly being used}
}
\value{
a vector of \code{n} length
}
\description{
use a vector repeatedly to generate a new vector of the specified length
}
\examples{
coerce_x_len(1, 10)

}
