% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosslink.R
\docType{class}
\name{CrossLink-class}
\alias{CrossLink-class}
\alias{CrossLink}
\title{The CrossLink class}
\description{
This Class is designed to manipulate coordinates of nodes with groups and to
visualize the crosslink network.
}
\details{
Nodes and Edges formed a network, and in the so called crosslink network,
Nodes in the same type will be placed in one column (named as a cross), and
Edges is the links linking between crosses.


cross: list names are inherited from cross.by in nodes; names' order (to control it using factor) is used as column order.

layout: list names are layout names. For each layout, a data.frame is stored
with node (key), node.type (node, top, bottom or header), x, y (coordinates), cross (which cross it belongs to).

canvas: deprecated

active: a layout name which is being used currently

params: other params used for internal functions
}
\section{Slots}{

\describe{
\item{\code{nodes}}{data.frame for node data}

\item{\code{edges}}{data.frame for edge data}

\item{\code{cross}}{a list, saving nodes in each cross, initialized by \code{\link{crosslink}}}

\item{\code{layout}}{a list, saving different layout of coordinates of nodes in crosses.}

\item{\code{active}}{character, Name of the active, or default, layout; settable using \code{\link{cl_active}}}

\item{\code{params}}{list.}
}}

