% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosslink.R
\name{crosslink}
\alias{crosslink}
\title{Generator of class CrossLink}
\usage{
crosslink(
  nodes,
  edges,
  cross.by,
  gaps = NA,
  spaces = NA,
  xrange = c(0, 10),
  yrange = c(0, 10),
  key.by = 1,
  src.by = 1,
  tar.by = 2,
  odd.rm = FALSE
)
}
\arguments{
\item{nodes}{a data frame with unique keys in the 1st (settable by \code{key.by}) column and metadata (used for aesthetic of nodes) in the other columns}

\item{edges}{a data frame with source and target nodes in the first two (settable by \code{src.by} and \code{tar.by}) columns and other metadata.}

\item{cross.by}{name of the column in \code{nodes} to be used to group nodes in different crosses. The order of crosses and nodes could be controlled by setting levels for corresponding columns in \code{nodes} using \code{\link{factor}} or \code{\link{forcats::fct_relevel}}}

\item{gaps}{a numeric vector of n-1 length, n is equal to length of crosses (that is \code{length(object@crosses)}).
The gaps vector is c(gaps between adjacent crosses).
set NA to use equally divided gaps.}

\item{spaces}{a named list of numeric vectors. The names must be names of crosses to be changed.
And each vector is of n+1 length, n is equal to length of nodes in their belonging cross.
The vector is c(top flank, spaces between adjacent nodes, bottom flank).
set NA to use equally divided spaces.}

\item{xrange, yrange}{canvas range}

\item{key.by}{name or index of the column in \code{nodes} to be used as key of nodes}

\item{src.by}{name or index of the column in \code{edges} to be used as source of links (edges)}

\item{tar.by}{name or index of the column in \code{edges} to be used as target of links (edges)}

\item{odd.rm}{remove odd nodes that are not included in edges (without links).}
}
\value{
a CrossLink object
}
\description{
Achieve a initialized CrossLink object by providing nodes, edges and cross.by.
}
\examples{

}
