% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{degree_to_radian}
\alias{degree_to_radian}
\title{convert degree to radian}
\usage{
degree_to_radian(degree)
}
\arguments{
\item{degree}{angle in degree}
}
\value{
angle in radian
}
\description{
convert degree to radian
}
\examples{
degree_to_radian(180)

}
\keyword{internal}
