% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gen_demo}
\alias{gen_demo}
\title{generate a toy CrossLink object}
\usage{
gen_demo(n_cross = 2, n_link = c(10), n_node = c(3, 4), seed = 6)
}
\arguments{
\item{n_cross}{the number of crosses (columns)}

\item{n_link}{a vector of numbers of links between crosses}

\item{n_node}{a vector of numbers of nodes in each cross}

\item{seed}{random seed for random links}
}
\value{
a CrossLink object
}
\description{
generate a toy CrossLink object
}
