% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosslink.R
\name{get_cross}
\alias{get_cross}
\title{Get cross data for plotting}
\usage{
get_cross(object, layout = NULL)
}
\arguments{
\item{object}{a CrossLink object}

\item{layout}{name of the layout to be retrived or set}
}
\value{
data.frame
}
\description{
Retrieve 2D coordinates of crosses
}
