% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{index_to_name}
\alias{index_to_name}
\title{a helper function to get names by integer indexes from data.frame or list}
\usage{
index_to_name(x, ind)
}
\arguments{
\item{x}{a data.frame or a list}

\item{ind}{a vector of integer indexes or colnames of \code{x}}
}
\value{
a vector of characters
}
\description{
a helper function to get names by integer indexes from data.frame or list
}
\examples{
index_to_name(data.frame(A = 1, B = 2), 1)

}
\keyword{internal}
