% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{layout_column}
\alias{layout_column}
\alias{layout_row}
\alias{layout_hive}
\alias{layout_polygon}
\alias{layout_arc}
\title{Predefined layouts}
\usage{
layout_column(
  object,
  layout_based = "default",
  layout_save = NULL,
  crosses = NULL
)

layout_row(
  object,
  layout_based = "default",
  layout_save = NULL,
  crosses = NULL
)

layout_hive(
  object,
  angles = NULL,
  crosses = NULL,
  layout_based = "default",
  layout_save = NULL
)

layout_polygon(
  object,
  angles = NULL,
  crosses = NULL,
  layout_based = "default",
  layout_save = NULL
)

layout_arc(
  object,
  angles = 30,
  crosses = NULL,
  layout_based = "default",
  layout_save = NULL
)
}
\arguments{
\item{object}{a CrossLink object}

\item{layout_based}{layout name to base on. Default: "default". This parameter is reserved for testing, and it is not recommended to change it!}

\item{layout_save}{layout name to save. set NA or NULL for active layout.}

\item{angles}{angles in degree for hive, polygon and arc. set NA or NULL for equally divided.}
}
\value{
a updated CrossLink object
}
\description{
Predefined layouts
}
\details{
\code{layout_column} is the default \code{default} layout.

\code{layout_row} and \code{layout_column} will always based on \code{default} layout.
}
