% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_along}
\alias{list_along}
\title{Create a list along given names and default values}
\usage{
list_along(names, default = NA, some = NA)
}
\arguments{
\item{names}{names of the list to be generated}

\item{default}{default value for each element of the generated list}

\item{some}{a list with elements whose names should be included in \code{names}, use this to set default values for some elements of the generated list}
}
\value{
a list
}
\description{
It can be useful to create an list with default values and some settable values. This is similar to the idea of seq_along(), which creates a vector of the same length as its input.
}
\examples{
list_along(names = c("A", "B", "C"), default = NA, some = list(B = 1))

}
