% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{merge_list}
\alias{merge_list}
\title{Merge two list}
\usage{
merge_list(x, y)
}
\arguments{
\item{x}{one list}

\item{y}{another list. Elements having same names in \code{x} will be override by those in \code{y}.}
}
\value{
a merged list
}
\description{
Merge two list
}
\examples{
merge_list(list(1, A = 2, B = 3, 4), list(A = 22, 5, 1))
}
