% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nullna_default}
\alias{nullna_default}
\title{a helper function to set default value when inputting NULL or NA}
\usage{
nullna_default(x, default)
}
\arguments{
\item{x}{input argument}

\item{default}{returned default value if x is NULL or NA}
}
\value{
x or default asis
}
\description{
a helper function to set default value when inputting NULL or NA
}
\examples{
nullna_default(NULL, c(1,2))

}
