% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{override_aes}
\alias{override_aes}
\title{Override default aes with new aes}
\usage{
override_aes(mapping = NULL, default_aes)
}
\arguments{
\item{default_aes}{aes with some default values}

\item{aes}{input aes}
}
\value{
a updated aes
}
\description{
Override default aes with new aes
}
\keyword{internal}
