% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{period_arcTrig}
\alias{period_arcTrig}
\title{get arc-trigonometric values in periods}
\usage{
period_arcTrig(
  x,
  y = NULL,
  arc.fun = c("acos", "asin", "atan", "atan2"),
  return.range = NULL
)
}
\arguments{
\item{x}{trigonometric values}

\item{y}{used for atan2 function. see \code{\link{atan2}}}

\item{arc.fun}{arc-trigonometric function}

\item{return.range}{set NULL for a range of 0 to 2*pi.}
}
\value{
angles in radian
}
\description{
get arc-trigonometric values in periods
}
\keyword{internal}
