% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{period_variable}
\alias{period_variable}
\title{get period variables with same return value}
\usage{
period_variable(
  x,
  period,
  range = c(0, period),
  range.inclusion = c("left", "both", "right", "neither"),
  simplify = F
)
}
\arguments{
\item{x}{a vector of variables in one period}

\item{period}{the period}

\item{range}{the range in which variables (usually with the same values with x) will be returned}

\item{range.inclusion}{left or right inclusion for \code{range}}

\item{simplify}{return list or not, see \code{\link{sapply}}}
}
\value{
a vector
}
\description{
get period variables with same return value
}
\examples{
period_variable(pi, 2*pi, c(-pi, 5*pi))

}
\keyword{internal}
