% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{Predefined_layouts}
\alias{Predefined_layouts}
\alias{layout_column}
\alias{layout_row}
\alias{layout_hive}
\alias{layout_polygon}
\alias{layout_arc}
\alias{layout_custom}
\title{Predefined layouts}
\usage{
layout_column(object, layout_save = "column")

layout_row(object, layout_save = "row")

layout_hive(
  object,
  angles = NULL,
  layout_based = "initial",
  layout_save = "hive"
)

layout_polygon(
  object,
  angles = NULL,
  layout_based = "initial",
  layout_save = "polygon"
)

layout_arc(object, angles = 30, layout_based = "initial", layout_save = "arc")

layout_custom(object, angles = 30, layout_save = "arc")
}
\arguments{
\item{object}{a CrossLink object}

\item{layout_save}{layout name to save}

\item{angles}{angles in degree for hive, polygon and arc. set NA or NULL for equally divided.}
}
\value{
a updated CrossLink object
}
\description{
Predefined layouts
}
\details{
\code{layout_column} is the default \code{initial} layout.

\code{layout_row} and \code{layout_column} will always based on \code{initial} layout.
}
\examples{

}
