% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{radian_to_degree}
\alias{radian_to_degree}
\title{convert radian to degree}
\usage{
radian_to_degree(radian)
}
\arguments{
\item{radian}{angle in radian}
}
\value{
angle in degree
}
\description{
convert radian to degree
}
\examples{
degree_to_radian(pi)

}
\keyword{internal}
