% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{segment_range}
\alias{segment_range}
\title{cut a segment defined by a range in 1D coordinate into n pieces}
\usage{
segment_range(range = c(0, 1), n, scale = NA, return.cutter = T)
}
\arguments{
\item{range}{a vector of 2 numeric values defining a segment}

\item{n}{cut the segment into \code{n} pieces}

\item{scale}{a vector of \code{n} numeric values to control segment sizes relatively. Set NA or NULL for equally dividing.}

\item{return.cutter}{only return the coordinate values of internal cutters}
}
\value{
a vector of n+1 or n-1 if \code{return.cutter}
}
\description{
cut a segment defined by a range in 1D coordinate into n pieces
}
\examples{
segment_range(c(0,10), 3, scale = c(1, 2, 7), return.cutter = F)

}
