% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosslink.R
\name{set_header}
\alias{set_header}
\title{Set Headers for each cross}
\usage{
set_header(
  object,
  header = NULL,
  hjust = 0.5,
  vjust = 0.95,
  hjust.by.nodes = F,
  vjust.by.nodes = F,
  layout = NULL
)
}
\arguments{
\item{object}{a CrossLink object}

\item{header}{a vector of titles for all crosses. Its length should be equal with the length of crosses (\code{object@cross}). Set NULL for default using \code{names(object@cross)} and set NA to remove headers.}

\item{hjust, vjust}{a vector for setting x or y position for each header. Repeatedly used if its length less than the length of crosses}

\item{hjust.by.nodes, vjust.by.nodes}{hjust and vjust will be setted based on nodes ranges (w/o blanks) or cross ranges (w/ blanks). Repeatedly used.}

\item{layout}{name of the layout to be retrived or set}
}
\value{
an updated CrossLink object
}
\description{
Set Headers for each cross
}
\examples{

}
