% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosslink.R
\name{set_space}
\alias{set_space}
\title{change spaces in default layout}
\usage{
set_space(object, spaces)
}
\arguments{
\item{object}{a CrossLink object}

\item{spaces}{A named list of numeric vectors or characters of "equal", "partition" and "flank" (see \code{\link{crosslink}}).
For each numeric vector, its length must be n+1, and n is the number of nodes in the corresponding cross;
The vector is c(top flank, spaces between adjacent nodes, bottom flank);
"equal", "partition" and "flank" will be converted to numeric vectors internally.
Only crosses whose names are among the list's names will be changed.}
}
\value{
an updated CrossLink object
}
\description{
change spaces in default layout
}
\details{
This function will only change default layout.
So, it is recommended to use this function after \code{\link{crosslink}},
and before transformation.
}
\examples{

}
