% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosslink.R
\name{show,CrossLink-method}
\alias{show,CrossLink-method}
\title{Print the CrossLink object}
\usage{
\S4method{show}{CrossLink}(object)
}
\arguments{
\item{object}{a CrossLink object}
}
\value{
No value is returned.
}
\description{
Print the CrossLink object
}
\keyword{internal}
