% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{string_to_circle}
\alias{string_to_circle}
\title{Get its center, radius and angle from a circle's string}
\usage{
string_to_circle(x, y, xend, yend, radian = NULL, radius = NULL)
}
\arguments{
\item{x, y, xend, yend}{start and end points defining a string of a circle}

\item{radian}{angle of the string.}

\item{radius}{radius of the circle. Only used if \code{radian} is NULL or 2*pi.}
}
\value{
a list of center, radius, string.start, string.end, angle, angle.start, angle.end
}
\description{
Get its center, radius and angle from a circle's string
}
\details{
For a string, given its radian or radius, there will be two circle
centers matched. This function will return both centers in a order of left
center and right center. For brevity, the radian will always less than pi (0-2*pi).
}
\examples{
string_to_circle(-1, 0, 1, 0, pi)
string_to_circle(-1, 0, 1, 0, pi/2)
string_to_circle(0, 1, 0, -1, pi/2)

}
