% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfromation.R
\name{tf_fun}
\alias{tf_fun}
\title{Transformation by a given function}
\usage{
tf_fun(
  object,
  fun,
  along = c("x", "y", "xy"),
  layout = "transforming",
  xrange.from = NULL,
  yrange.from = NULL,
  xrange.to = NULL,
  yrange.to = NULL,
  crosses = NULL,
  ...
)
}
\arguments{
\item{object}{a CrossLink object}

\item{fun}{function for coordinate transformation}

\item{along}{another axis will changed along this given axis. Such as, if along = "x", transformation function will change y values along x axis and keep x values asis.}

\item{layout}{layout name to save}

\item{xrange.from, yrange.from}{the range to be scaled from before transformation by function}

\item{xrange.to, yrange.to}{the range to be scaled to}

\item{crosses}{only these crosses will be transformed}

\item{...}{other parameters passed to \code{fun}}
}
\value{
an updated CrossLink object
}
\description{
Transformation by a given function
}
\examples{

}
