% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{todataframe_by_list}
\alias{todataframe_by_list}
\title{convert a list into a data.frame}
\usage{
todataframe_by_list(list)
}
\arguments{
\item{list}{a list with equal length of its each element}
}
\value{
data.frame
}
\description{
convert list (a list with equal length of each element) into a data.frame
}
\examples{
todataframe_by_list(list(num = c(1:10), cat = c(rep("A", 5), rep("B", 5))))

}
\keyword{internal}
