% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tolist_by_group}
\alias{tolist_by_group}
\title{convert a vector into a list}
\usage{
tolist_by_group(x, group, drop = F)
}
\arguments{
\item{x}{a vector}

\item{group}{a vector with equal length of x, providing group information of elements in x}

\item{drop}{drop levels of x}
}
\value{
a list. Order of elements is controled by levels of group.
}
\description{
convert x (a vector) into a list according to the group information (another vector) of elements in x.
}
\examples{
tolist_by_group(x = 1:10, group = c(rep("A", 5), rep("B", 5)))

}
\keyword{internal}
