% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfromation.R
\name{transform_by_fun}
\alias{transform_by_fun}
\title{Transform 2D coordinates by a function(x, y)}
\usage{
transform_by_fun(
  x = NULL,
  y = NULL,
  fun,
  along = c("x", "y", "xy"),
  xrange.to = NULL,
  xrange.from = NULL,
  yrange.to = NULL,
  yrange.from = NULL,
  ...
)
}
\arguments{
\item{x}{a vector of coordinates at x axis}

\item{y}{a vector of coordinates at y axis}

\item{fun}{a function taking \code{x} and/or \code{y} as arguments}

\item{along}{th independent variables of \code{fun}}

\item{xrange.to}{before transforming, rescale \code{x} from \code{xrange.from} into the range of \code{xrange.to}, set NULL to use \code{range(x)}}

\item{xrange.from}{range of \code{x}, set NULL to use \code{range(x)}}

\item{yrange.to}{same with \code{xrange.to}, but for y axis}

\item{yrange.from}{same with \code{xrange.from}, but for y axis}

\item{...}{other parameters passed to \code{fun}}
}
\value{
a matrix of transformed coordinates
}
\description{
Transform 2D coordinates by a function(x, y)
}
\examples{
xy <- data.frame(x = 1,
                 y = 1:100)
xy_tf <- transform_by_fun(xy$x, xy$y, function(x) 0.1*sin(x), along = "y", yrange.to = c(0, 2*pi))
plot(xy)
plot(xy_tf, type = "o")

}
