% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfromation.R
\name{transform_matrix}
\alias{transform_matrix}
\title{a helper function for 3*3 matrix}
\usage{
transform_matrix(a = 1, b = 0, c = 0, d = 0, e = 1, f = 0, g = 0, h = 0, i = 1)
}
\arguments{
\item{a}{1st value of 3*3 matrix}

\item{b}{2nd value of 3*3 matrix}

\item{c}{3rd value of 3*3 matrix}

\item{d}{4th value of 3*3 matrix}

\item{e}{5th value of 3*3 matrix}

\item{f}{6th value of 3*3 matrix}

\item{g}{7th value of 3*3 matrix}

\item{h}{8th value of 3*3 matrix}

\item{i}{9th value of 3*3 matrix}
}
\value{
a 3*3 matrix as follows:
matrix(
a, b, c
d, e, f
g, h, i
)
}
\description{
a helper function for 3*3 matrix
}
\examples{
transform_matrix(a = 2, e = 2, i = 2)

}
