% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfromation.R
\name{transform_matrix_affine}
\alias{transform_matrix_affine}
\title{generate a 3*3 matrix for affine transformation}
\usage{
transform_matrix_affine(
  type = c("none", "translate", "scale", "rotate", "shear", "reflect"),
  x = 0,
  y = 0,
  theta = NA,
  scale = NULL,
  counterclockwise = FALSE
)
}
\arguments{
\item{type}{a affine transformation type}

\item{x}{transform value along x axis}

\item{y}{transform value along y axis}

\item{theta}{transform value of angle in radian, usually in 0 to 2pi}

\item{scale}{scale size, only used for \code{scale} type}

\item{counterclockwise}{control directions of theta or x,y if type is "shear"}
}
\value{
a 3*3 matrix
}
\description{
\subsection{learn more about 2D affine and perspective tranformation matrix:}{
\itemize{
\item https://en.wikipedia.org/wiki/Transformation_matrix
\item https://www.cnblogs.com/bnuvincent/p/6691189.html
\item \code{Rconic::`Affine planar transformations matrix`}
}
}

\subsection{Used arguments for different types:}{
\itemize{
\item none: ignore everything
\item translate: x, y, counterclockwise
\item scale: x, y, theta, scale
\item rotate: x, y, theta (angle of the rotation), counterclockwise
\item shear: x, y, counterclockwise
\item reflect: x, y, theta (angle from x axis), counterclockwise
}
}
}
\examples{

# eigen matrix
transform_matrix_affine(type = "none")

# shift matrix for convert c(0, 0) to c(1, 2)
mat <- transform_matrix_affine(type = "translate", x = 1, y = 2)
mat \%*\% c(0, 0, 1)

# rotation matrix for convert c(0, 0) to c(0, 2)
mat <- transform_matrix_affine(type = "rotate", x = 1, y = 1, theta = pi/2)
mat \%*\% c(0, 0, 1)

# shear for convert c(1, 1) to c(2, 1) if along x or c(1, 2) if along y
mat <- transform_matrix_affine(type = "shear", x = pi*45/180, y = 0)
mat <- transform_matrix_affine(type = "shear", x = 0, y = pi*45/180, counterclockwise = T)

# reflect for convert c(0, 0) to c(2, 2)
mat <- transform_matrix_affine(type = "reflect", x = 1, y = 1, theta = pi*45/180)

}
\keyword{internal}
