% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vector_to_coord}
\alias{vector_to_coord}
\title{convert a vector of characters to a vector of numeric coordinates}
\usage{
vector_to_coord(x, scale = NA, range = c(0, 1))
}
\arguments{
\item{x}{a vector of characters, which will be coerced into factor.}

\item{scale}{a vector of \code{length(x) - 1} numeric values to control adjacent spacing. Set NA or NULL for equally dividing.}

\item{range}{a vector of 2 numeric values defining a interval, where \code{x} will be mapped.}
}
\value{
a vector of length(x)
}
\description{
convert a vector of characters to a vector of numeric coordinates
}
\examples{
vector_to_coord(1:4, scale = c(2,2,2), range(1, 9))

}
\keyword{internal}
