% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cl_void}
\alias{cl_void}
\title{a theme for crosslink plot}
\usage{
cl_void(gg)
}
\arguments{
\item{gg}{a ggplot object}
}
\value{
an updated ggplot object
}
\description{
a theme for crosslink plot
}
\examples{

}
