% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfromation.R
\name{tf_affine}
\alias{tf_affine}
\alias{tf_rotate}
\alias{tf_flip}
\alias{tf_shift}
\alias{tf_shear}
\alias{tf_scale}
\title{Affine Transformation}
\usage{
tf_affine(
  object,
  type = c("none", "translate", "scale", "rotate", "shear", "reflect"),
  by.each.cross = F,
  x = 0,
  y = 0,
  angle = NA,
  scale = NA,
  counterclockwise = FALSE,
  relative = T,
  crosses = NULL,
  layout = "transforming"
)

tf_rotate(object, x = 0, y = 0, angle = 90, ...)

tf_flip(object, axis = c("x", "y"), ...)

tf_shift(object, x = 0, y = 0, ...)

tf_shear(object, axis = c("x", "y"), angle = 15, ...)

tf_scale(object, x = 0, y = 0, angle = 90, scale = 1, ...)
}
\arguments{
\item{object}{a CrossLink object}

\item{type}{affine type}

\item{by.each.cross}{transformation to be performed by each cross}

\item{x, y}{scale about this position}

\item{angle}{transformation angle in degree}

\item{scale}{scale size for type \code{scale}}

\item{counterclockwise}{angle in counterclockwise for transformation}

\item{relative}{x,y is relative coordinates in range c(0,1)}

\item{crosses}{only these crosses will be transformed}

\item{layout}{layout name to save}

\item{...}{other parameters passed to tf_affine}

\item{axis}{along which axis to flip or shear}
}
\value{
an updated CrossLink object
}
\description{
Wrapper functions for transformations,
rotate: \code{tf_rotate}, flip: \code{tf_flip}, shift: \code{tf_shift} shear: \code{tf_shear}, scale: \code{tf_scale}
}
\details{
\subsection{learn more about 2D affine and perspective tranformation matrix:}{
\itemize{
\item https://en.wikipedia.org/wiki/Transformation_matrix
\item https://www.cnblogs.com/bnuvincent/p/6691189.html
\item \code{Rconic::`Affine planar transformations matrix`}
}
}

\subsection{Used arguments for different types:}{
\itemize{
\item none: ignore everything
\item translate: x, y, counterclockwise
\item scale: x, y, theta, scale (scale about x,y point, scale's unit size is \code{scale*cos(theta), scale*sin(theta)})
\item rotate: x, y, theta (angle of the rotation), counterclockwise
\item shear: x, y, counterclockwise
\item reflect: x, y, theta (angle from x axis), counterclockwise
}
}
}
\examples{


}
